feat: write actions to select and unselect items
