+++ tsconfig.app.json
@@ -2,9 +2,9 @@
   "compilerOptions": {
     "strict": true,
     "esModuleInterop": true,
-    "lib": ["dom", "es2015"],
+    "lib": ["dom", "es2016"],
     "jsx": "react-jsx",
-    "target": "es5",
+    "target": "es6",
     "allowJs": true,
     "skipLibCheck": true,
     "allowSyntheticDefaultImports": true,
